Add a VulnerabilityAnalysis REST API endpoint #103
