+++ dejacode/urls.py
@@ -51,6 +51,7 @@ from product_portfolio.api import ProductComponentViewSet
 from product_portfolio.api import ProductDependencyViewSet
 from product_portfolio.api import ProductPackageViewSet
 from product_portfolio.api import ProductViewSet
+from product_portfolio.api import VulnerabilityAnalysisViewSet
 from reporting.api import ReportViewSet
 from vulnerabilities.api import VulnerabilityViewSet
 from workflow.api import RequestTemplateViewSet
@@ -80,6 +81,7 @@ api_router.register("reports", ReportViewSet)
 api_router.register("external_references", ExternalReferenceViewSet)
 api_router.register("usage_policies", UsagePolicyViewSet)
 api_router.register("vulnerabilities", VulnerabilityViewSet)
+api_router.register("vulnerability_analyses", VulnerabilityAnalysisViewSet)
 
 
 urlpatterns = [
